Bug 657991 — Use asynchronous property setters
Port the contact pane to use the new asynchronous property setters in folks.
This isn't as tidy as it should be: ideally each editing widget constructor
would take an async delegate which it calls to set the appropriate property
values in folks. However, async. delegates aren't supported yet by Vala
(see: bgo#604827). Therefore we have to make do with a big switch statement
on the property names. :-(

Closes: bgo#657991