Added undo for some link opeartions
Added undo for link when happening through inline suggestions.
Added undo when happening through link-to-my-contacts mode of the LinkDialog.

Tthere's other possible approach, maybe prettier, it's passing the names of the linked contacts, needed for the notification, inside the LinkOperation object. That will need some judgement of the ancients, then will do.

https://bugzilla.gnome.org/show_bug.cgi?id=668751