Try to left align avatar in content pane
Its hard to left align the avatar in the content pane as its inside
a button with theme-dependent size. We hack around it by creating
a special row group for that row and setting no spacing
after column 0.

This doesn't quite get it right, but it looks better at least.