Adding webcam support when setting new avatars.
I think the implementation is kinda ugly because of those all if/endif.
But I can think of other way of doing it.
So far the testing for a webcam device is kinda hacky, but it works
without too much hazard.
The other detection method I can use is the one in cheese and that
includes using gudev, and I didn't want to add a new dependency.