Contacts: fix changing a Google Other contact into a My Contact
Previously, clicking on the Add to My Contacts button for a Google
Other contact would attempt a link operation with nothing, that
would end up creating a new scratch contact to link into, which
by chance would also have the magic in-google-personal-group bit set.
Instead, use the new API in Folks to change it directly, to avoid
creating a duplicate contact.

https://bugzilla.gnome.org/show_bug.cgi?id=687049