Contact: Make keep_widget_uptodate callback owned
Without this the callback data will not get reffed when this is called
so it may later point to uninitialized memory. This happened to me
when deleting a contact and then undoing that, since ContactSheet.update()
had exited so the closure was freed but the keep_uptodate stuff
still referenced it.