Sync cheese-flash.[ch]
And properly modify callers to the new API.

Old code was creating a window with the size of the screen. Thus,
when taking a photo, the gnome-shell panel was then "hidden" by
this big window. This has been fixed in "upstream" cheese, so, grab
the latest code and make small modifications to fit the new API.

https://bugzilla.gnome.org/show_bug.cgi?id=702177