ContactEditor: Move string constants from Contact
This fixes a crash when creating a new contact in an empty address book,
caused by the string arrays only being populated when the Contact class
is initialised — which doesn’t happen if the address book is empty.

These arrays were only used in the ContactEditor class, so they
shouldn’t have been in the Contact class in the first place.

https://bugzilla.gnome.org/show_bug.cgi?id=735168