Show address fields when adding the first contact
The new contact dialog iterates over the list of address fields, which
are declared as static in the Contact class. Static data in Vala is
implemented by initializing the data during class_init(), or in other
words, the first time that the class is instantiated. When adding the
first contact, there are no instances of Contact, and so the address
fields have not been initialized. This leads to a blank space in the new
contact dialog, where the list of address fields should be.

Ensure that class_init() of the Contact class has been called by
calling g_type_class_ref() on the Type before iterating over the address
fields.

https://bugzilla.gnome.org/show_bug.cgi?id=702810