Make labels in contact-sheet selectable
Right now to be able to copy & paste a name, address or telephone
number one has to change to edit mode, select the text and then
copy it. This is not a good UI, therefor thes patch makes the labels
in the contact-sheet selectable. This way the user doesn't have to
change to edit mode to be able to copy & paste the information.

https://bugzilla.gnome.org/show_bug.cgi?id=710485