SearchProvider: improve performance and add description.
Note that we don't need all the code from our own custom classes such as
Contacts.Contact, we only want speed. The best way to do this is
to put away all unnecessary operations and to just keep it simple.

We're falling back to the SearchView of Folks. That means that the only
way of improving performance, is to make SearchView better for this
case.