Store: remove the changed signal.
We only use this for the find_contact method, which on its own is only
called when the program is started. As such, the likelihood that a
contact changed between startup and at the end of searching is
negligible. Removing this also takes away some overhead in both
performance and memory usage.