Revert "Store: add contacts in a separate thread."
This reverts commit 9c256a5c7917a01f844c43501468c4700a4fcc28.

Tried adding multithreading a bit too naively at that point. It's
possible to get it working using a Glib.RWLock, but that makes the code
whole lot less clean, so I don't want to use that. In the long term, we
need to think of a good strategy for this problem (such as possibly not
having a separate arraylist for the Contacts anymore).