WIP: Rewrite the ContactEditor/ContactSheet
Rather than using a large if-else construct with several copy-pasted
chunks, we create a `PropertyField` class, which represents a property
instance of a persona. These object can then create an appropriate
widget that allowss the user to view, edit or remove that specific
property.

As a consequence, we can extract and simplify parts of the code on how
to keep the list of properties saved: we can just have a `GListModel` of
said property instances in the common `ContactForm` parent class. The
subclasses `ContactSheet` and `ContactEditor` can then deal with this
list appropriately.

In the long term, we can simplify this probably even more; for now, it's
good enough to have this over the old code, which was getting a bit
unmaintainable.