Store: remove eds_persona_store_changed()
Due to a race condition, a persona store could be available _before_ the
signal was handled. That meant that sometimes this signal wasn't
triggered when it really should've been. The easy fix (and imho) cleaner
solution is to not provide any E-D-S specific behaviour, and to just
listen to newly available backends in general.

Fixes https://gitlab.gnome.org/GNOME/gnome-contacts/issues/114.