Typeset: change how we handle VcardTypeMapping
This removes the generic and unclear name `InitData` in favor of
`VcardTypeMapping`, which is intuitively much more clear on what it
actually is used for.

Instead of building a hash table and adding each type as a key, we just
use a normal list. This won't add too much overhead and makes the data
structure much less redundant (and easier to work with) and obscure.