TypeCombo: simplify and document
Remove all the code for a custom user-defined label (using an entry).
This code path has been dead for a while, and it's better to use the
built-in entry of a ComboBox anyway. This way, we can alo reintroduce it
later.

We can now also derive directly from a Gtk.ComboBox, which simplifies
the code even further.

Finally, this commit adds some documentation in the TypeCmbo, which
hopefully makes it easier to understand for new contributors.