ContactEditor: Don't show the "Linked Accounts" button
By executing `show_all()` on the whole ContactEditor, we show elements
again that we explicitly set to be invisible -for example the "Linked
Accounts" button which should not be shown for newly created contacts.

We should avoid show_all() as much as possible, but to refrain from
having large changes this late in the schedule, let's just call the
`show_all()` on the container grid, which does not have elements that
shouldn't be shown.

Reported in https://gitlab.gnome.org/GNOME/gnome-contacts/issues/124