Use individual directly
This removes most of the Contact class and makes use of the Individual
of folks directly.
The aim of this refactor is to make the code more readable.
Since Individual provides most of the needed features the Contact class
adds a complicated abstraction without any real benefit.

Transform FakePersona a real Persona which can be used as a placeholder
till folks returns the real Persona.

Use SearchQuery to find the best matching contact when opening
gnome-contacts with a e-mail address