Editor: use listbox layout to edit contact and secondary menu
GNOME uses now listboxes as the standart design pattern instead of a
grid. This replaces the grid and makes use of listboxes to allow the
user to edit a contact.
Some key features are:
 - Hide less important properties when not used
 - Dynamically fill the editor with properties so that the user has always
   one empty row to fill for each visible property
 - use a dialog for the birthday picker
 - Group properties by persona

ContactSheet:
Replace the edit button with a secondary menu.
The secondary menu contains share (hidden for now), edit, unlink and delete.
The reason for this change is that it doesn't make a lot of sense to have
delete and unlink inside the edit mode, since they don't require to commit changed.

Folks doesn't provied a staging features. So changes are commited
directly to the backend. The FakePersona and FakeIndividual are used
exactly for this. They work as a intermidiate layer so the editor can
change the persona directly and then when the user presses "done" the
changes can be copied to the real contact.

This cleans up also how linking and unlinking is done, by using the methodes
provided by folks.