flatpak: Disable BlueZ backend
We figured out in libfolks that we were accidentally not properly
requiring the `python3-dbusmock` module that is necessary for the BlueZ
tests. We might want to be smarter about this in the future, but let's
just lose the offending backend for now. (Note that the flatpak manifest
doesn't even allow us to connect to the specified D-Bus interface).