Make Cheese a feature dependency
libcanberra-gtk got removed in the latest Flatpak runtime (3.38), so we
can't build cheese as it hard depends on it. Luckily, we don't
hard-depend on cheese, so let's get rid of it (for now).

As an extra, this commit also makes Cheese an optional dependency,
making it a bit easier to build it at the first go (without installing
cheese-devel) and a shorter CI time.