app: Don't add accelerators before startup()
For certain accelerators (like `<Primary>`), GTK needs to be able to
check the `GdkDisplay` under which it is running. Since this first needs
`gtk_init()` to be called, we need to wait until the
`GLib.Application.activate()` signal, where this call is done for us.

Fixes https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/169