App: Cleanup the window creation code
The current code that created a new window was a mess, and didn't really
say what or why we were doing things. This commit cleans up some of this
by renaming some variables and functions and making sure we always show
a window to the user, so that we can provide useful feedback.