ui: Don't set SetupWindow's visible property
When initializing a widget, the template is processed before the
construct properties. That's usually fine, but showing a toplevel
window has the side effect of realizing it; doing that before the
window is associated with an application means that properties like
the GTK application ID won't be exported, and gnome-shell fails to
match the window to the correct .desktop file as a result.

Just remove the property from the template, the code already handles
showingi the window.

See https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/2893