MainWindow: Rework window state restore
The current mechanism for restoring the MainWindow state used deprecated
API and did a lot of unnecessary work (keeping the window state in sync
on each window resize, rather than doing it only once). This commit
improves that by not using the deprecated GdkScreen API anymore and by
only saving the window state on the destroy signal.