contacts-main-window: cancelling selection no longer hides headerbar buttons
As described in #184, selecting a contact and then exiting selection mode
by pressing the cancel button causes the favorite and menu buttons to disappear.

To fix this, the make  cancel button's callback change the UiState to SHOWING
rather than NORMAL if a contact is currently selected. The headerbar buttons
then remain displayed as expected.

Closes #184