contact-sheet: stop cursor from appearing in contact pane values
As described in #199, clicking anywhere on a value in the selected contact pane
(such as the contact's name or email address) causes a cursor to appear
after the character which was clicked, making the value look editable.

Set the name and row value labels' can_focus attribute, to false
stopping the cursor from appearing (as expected).

Fixes #199