data: Temporarily disable appstream validation
It seems some builds of appstream-util don't like the GNOME 40
versioning scheme if the development releases are part of the appdata.
This will have to get fixed in appstream-glib, but as long as we don't
have that, let's just disable the test for now.

For the upstream appstream-glib issue, see
https://github.com/hughsie/appstream-glib/issues/391