main-window: fix false title after cancelling selecting
When you select more than one contact and than click "Cancel",
the title in the left headerbar is "1 Selected".

This is very irritating and the title
should be instead again "Contacts".

The solution is to write "Contacts", when 0 are selected.
Before it does nothing in this case and the last title remains.

https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/204