Ensure correct handling of February 2000 in birthday editor
There is a third rule for leap year handling which is that years
divisible by 400 are leap years. Normally this isn't a huge issue since
it only happens every 400 years, but the year 2000 was one of those
years which is well within the reasonable range of current birthdays and
the current system will not accept the 29th of February 2000 as a valid
date.