Port to GTK4 and libadwaita
This is a mega-commit which ports Contacts to GTK4 and libadwaita, the
library which provides GNOME-specific widgets on top of GTK4.

This change also now follows the new mockups of Contacts, which use a
boxed list style to convey contact information.

There is a minor set of known issues which we'll still need to solve
later (preferably):

* For now, taking a picture with your webcam is not implemented. In
  GTK3, we used to do this with Cheese, but this hasn't been ported to
  GTK4. Ideally, we could just directly use Pipewire though.
* Some CRITICALs when we have some unexpectedly long names or property
  values
* The delete button is gone for most properties. This probably needs to
  be rethough at the design level on how we want to deal with it.
* We're still blocked a bit on libedataserverui not having a GTK4 port
  yet.