sheet: Add button to open a contact's address
This commit adds the ability for the user to open up GNOME Maps (or any
application that registers itself as a URI handler for "maps:" URIs) to
the specified location.

We do this by creating a "maps:" URI which takes the (encoded form of
the) address as a parameter. If there is one result, the maps
application can show it; if there are more, they can provide the user
with a list of selections.

Successfully tested that this works with GNOME Maps' nightly flatpak.

See also: https://gitlab.gnome.org/GNOME/gnome-maps/-/merge_requests/199

Fixes: https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/87