Support showing a role of a contact
The role property contains 2 (optional) elements: an organisation name
(note that "organisation" is broader than "company"), and the role of
the person in that organisation (for example: "Board Member").

This is mostly useful for people who work in a corporate setting, but
can also be nice to keep track of what your relatives are doing.

In the ContactSheet (ie. viewing mode), the property shows itself as a
single row: "$ROLE at $ORGANISATION"; the text may slightly differ if
one of the two is not available. In the ContactEditor, it gets split up
into two rows in the same listbox: one for the organisation and one for
the actual role.

Finally, note that a contact can have multiple roles/organisations, so
this property can occur more than once.