Use a ListModel for address books (PersonaStore-s)
Remove the util function `get_eds_address_books()` and the several copy
pastes that deal with `Folks.Backend`s coming up and dynamically
changing the list of address book.

Instead, `Contacts.Store` now exposes a `GLib.ListModel` that something
like the `Contacts.AccountsList` widget can use to generate a
`Gtk.ListBox`.

While we're at it, remove `AddressBookList`, as it's not used anywhere.

Fixes a problem in `Contacts.AccountsList` where it showed an address
book as selected, but it didn't seem to register it properly.

Fixes: https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/238