sheet: Add a button to start a phone call
evolution-data-server provides a class `E.PhoneNumber` which -amongst
other things- allows us to parse freeform strings as phone numbers. If
successful, it also allows to generate a `tel:` URI which we can use to
start a call with the user's preferred application. Note however that
it's only really supported if E-D-S was compiled with support for
libphonenumber (which provides the underlying implementation).