query-filter: Optimize for empty SimpleQuery
When starting the application (or when the user removes the input from
the search entry), the `QueryFilter` object we use to filter our
contacts will match any item. GTK allows us to specify this in the
`get_strictness()` method by returning `Gtk.FilterMatch.ALL`, which
is a performance improvement, as it doesn't have to iterate over all the
items in the list anymore.