Contact: allow the contacts store to be null
We need to know about the primary store in a `Contacts.Contact` for 2
reasons:

1. To find out which persona is the "primary" persona (if there is one)
2. To save any new properties (which are not linked to a persona) in the
   primary store

However, in some cases we do not have a special need for these 2
properties: in unit tests, and when we want to port our
(de-)serialization code to work with an array of `Contacts.Contact`
objects (especially with the importer being in a separate process).

In those cases, it should be allowed to pass `null` instead.