contact: Avoid ChunkFilter for finding chunks
The ChunkFilter class is an easy way of filtering out properties.
However, it's not part of "core", so we probably shouldn't using it from
within `Contacts.Contact`. This is also a slight performance
optimization since we can avoid some of the overhead from using a
`GtkFilterListModel`, as well as some unnecessary refcounting.