contact: Don't keep a reference to the store
First, we're only using the `Contacts.Store` to get to the aggregator's
primary store, which we can just get from the
`Folks.PersonaStore.is_primary_store` property itself.

Second, this means we can make the constructor for a `Contact` a bit
simpler, as we don't have to actually pass on a `Contacts.Store`
reference.

Third, this parameterizes `apply_changes()` so that it can choose which
PersonaStore to save new details into, which is a step towards being
able to configure the address book when you're editing/creating a
contact.