utils: Ensure pure key-file individuals are ignored
Since linking is done by creating entries in the `relationships.ini`
file which end up as a Persona, we can have an individual which consists
of multiple Personas, all from the key-file backend. This is not
something we accounted for in the `Utils.is_ignorable` function, so this
commit fixes that.