contact: Copy the chunks before applying changes
When applying the changes of certain fields, we've seen that this leads
to a `individuals_changed_detailed()` being called with the same
individual in the `removed` and `added` set. The signal callback
propagates to several layers, until it lands in the
`Contact:on_individual_personas_changed()` function. There, all chunks
related to the persona are removed, even when we still might be applying
changes of some of the other chunks.

The `apply_changes()` method in other words should keep its own copy to
prevent that.

Fixes: https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/271