contact: Add QR code to share individual contacts
Sharing contacts in an easy and offline way is currently not possible.
Most mobile phones have a camera and are capable of scanning QR codes.
The vCard format is widely used to easily exchange contact information.
A contact can be saved in vCard format into a QR code.

A button with a QR code icon is added next to the "favourite" and "edit"
buttons. When the user presses this button, a dialog opens up, which
shows a QR code containing the current contacts data in vCard format.