hide start title buttons in expanded leaflet state
As the start title buttons are only supposed to be visible in the left-most header bar,
just as the end title buttons are in the right-most, every (i.e. the) other header bar has to not include them.

Here it depends on the leaflet foldedness state of the content box if there are two header bars,
so the title button property needs to be bound to this flag.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>