tests: Improve and add vCard tests
Add a few more vCard tests to make sure we're not regressing on our
vCard import feature.

One of the design desicions at the time was to make the
`Contacts.Io.Parser` take a `GLib.InputStream` so we can make it
agnostic of what kind of input was given. We can now use this to make
the vCard tests a bit simpler in general by using a constant string
inside the test file rather than having to deal with environment
variables to pass on a path.