main-window: replace actionbar with custom widget
The actionbar uses a centerbox and as such makes sure,
that the center widget is always centered if possible.
Therefore, the minimal width of it is twice the width
of the first two buttons (or the last one, whatever is
larger).

This actionbar, however, does not have any child in the
center widget. Therefore, depending on the language,
the sidebar can grow due to empty space in an actionbar.

Replicating the actionbar's behaviour regarding alignment
at start and end and spacing of its children without its
center widget avoids this large empty space.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>