avatar-selector: add file format filter to file dialog
Originally, the file dialog does not provide any filter.
Therefore, it is harder to find a file for an avatar, and
easier to select a non-image or an image with an
unsupported image format that would be not suitable here.

Hence, a format filter for any supported type as the
default makes it easier to make an expected input. Also,
adding each supported format individually helps users to
find an image in cases of directories with many image
files.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>