window: Confirm before navigating in edit mode
This fixes a bug where the application would enter a strange state after
the user opens a new contact through the search provider while editing
a different contact. Instead, we will add a new method,
`Contacts.MainWindow.navigation_request`, and only change the selected
contact if the result is true.

Fixes: https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/319