chunk: Add display-name and icon-name property
Rather than building a separate list of structs somewhere to fetch a
user-facing string/icon, let's put that knowledge directly into the
`Contacts.Chunk` object, since that's what we're working with anyway.

This also allows us to easily set an accesible label. As a consequence,
when using a screen reader, it will now actually tell you what field
you're focusing on when showing a contact.