app: Build AboutWindow from AppStream data
Use libadwaita's `Adw.AboutWindow.from_appdata()` constructor to build
the about window, filling in most details from the appstream data XML
file.

This change means we'll no longer have to keep the metadata in 2
separate places, and allows us to show the release updates from the
about window as well.

Note that we have to do a bit of awkward version juggling as the
AppStream specification does not deal with alphabetical components
inside a version (so we use $VERSION~alpha instead of $VERSION.alpha).