sheet: Use custom Gtk.ListBoxRow subclasses
Rather than using `Adw.ActionRow`s, this commit splits out the
`ContactSheetRow` helper class into a separate file, and changes it to
use a GTK template UI file. An even bigger change is that it no longer
extends `AdwActionRow` but instead implements a custom `GtkListBoxRow`.

By implementing a custom row, we can add extra API that might be too
niche of a use case for `AdwActionRow` to expose, such as setting the
text direction on the title of a row.