vcard: Add birthday tests with different time zones
It's happened before that Contacts showed the wrong date due to some
(implicit) time zone conversions, so let's make sure we test this.

While we're at it, extract some of the logic in parsing a single contact
birthday in a separate helper function so we don't need have too much
copy-pasta.