Allow taking a picture with the camera
We "lost" this feature with the GTK4 port. Re-implement this again with
the camera portal and a GStreamer pipeline to handle the pipewire
buffers.

Since we use the "gtk4paintablesink" element for integrating the
GStreamer's buffers with GTK4, we need to make sure that we have the
necessary dependencies in the Flatpak manifest too (since it's written
in Rust). For now, we solve this by copying Showtime's flatpak manifest
elements.

Thanks to Jordan Petridis and Maximiliano Sandova for figuring out the
Flatpak Rust integration.

Fixes: https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/219