flatpak: Disable some irrelevant g-o-a options
The gnome-online-accounts (g-o-a) build is failing in the CI due to some
missing dependency (libkeyutils) for the Fedora backend. Since the
Fedora backend doesn't actually matter for GNOME Contacts, we can just
fix the problem by not building it.

While we're at it, we can also turn off some other compile options that
aren't needed either for Contacts.