main-window: Make sure to rename "cancel" action eveyrwhere
In commit 3e84850def, we grouped together the several "cancel-*" actions
into a single "cancel" action. The idea being that we want to make sure
the "wrong" cancel method isn't being called.

However, in doing so, we forgot to rename a couple of instances,
effectively breaking the "Cancel" button in some cases in the UI. This
commit now pushes through the rename in those cases too.

Fixes: 3e84850def ("main-window: Allow Escape to cancel selection mode")