contact-editor-property: Add a crossfade animation
This chains up a crossfade animation to the slide down animation of the
contact editor property revealer. This will help showing the overflow of
the revealer without having a property overlaping another nor having its
shadow cut at the transition's border. This in turn will help fixing the
property spacing.

Signed-off-by: Adrien Plazas <aplazas@gnome.org>