Use glycin instead of GdkPixbuf for image loading
`libglycin` is a new library that sandboxes image loaders, and is
written in Rust for better memory safety. Let's reply good old
`GdkPixbuf`, which is showing its age at this point.

Since we rely on some newer APIs for being able to create a `GlyLoader`
out of a `GInputStream`, we need at least glycin version 2.0.alpha.